Decompose oversized tabletop v2 Svelte components
Split the five v2 tabletop components that breached the 150-line
Single Responsibility rule into focused base + sub-components. Fixes #93.

ZoneRenderer (646) is now a slim base renderer that composes ZoneFrame,
ZoneHeaderTab, ZoneEditToolbar (with ZoneTypeSwitch + per-type
ZoneSpreadSettings/ZoneGridSettings), ZoneGridLines, ZoneInsertIndicator
and ZoneResizeHandles — resolving the "base + per-type sub-components"
and "prefer snippets/components over if-cascades" violations.

TabletopV2 (522) is now a thin shell; its keyboard shortcuts, context-menu
state and canvas/viewport/drag wiring moved into co-located modules
(tabletopKeyboard, tabletopMenus, tabletopCanvas) with TabletopCanvasView,
TabletopContextMenus and TabletopSurface presentational components, per the
"business logic in .svelte.ts modules" guideline.

ComponentSidebar (320) splits into SidebarHeader, SidebarComponentItem and
SidebarComponentDetails. PileRenderer (263) and ShuffleAnimation (162) share
a new FlipCard and split out PileFace and PileCountBadge.

Every resulting component is under 150 lines. svelte-check passes with no
errors; the 332 tabletop v2 unit tests are unaffected.
